Road_Workspace/yolo-ml-project/requirements.txt
tensorflow>=2.4.0
numpy>=1.19.0
opencv-python>=4.5.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
PyYAML>=5.4.0